requirements.txt
amqp==1.4.9
anyjson==0.3.3
argon2==0.1.10
argon2-cffi==18.3.0
atomicwrites==1.2.1
attrs==18.2.0
billiard==3.3.0.23
boto3==1.9.61
botocore==1.12.61
celery==3.1.18
cffi==1.11.5
coverage==4.5.2
Django==2.1.3
django-bcrypt==0.9.2
django-environ==0.4.5
django-filter==2.0.0
django-nose==1.4.6
django-storages==1.7.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docutils==0.14
flake8==3.6.0
itsdangerous==1.1.0
jmespath==0.9.3
kombu==3.0.37
Markdown==3.0.1
mccabe==0.6.1
more-itertools==4.3.0
nose==1.3.7
pluggy==0.8.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py==1.7.0
py-bcrypt==0.4
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
PyJWT==1.7.0
pytest==4.0.1
pytest-cov==2.6.0
python-dateutil==2.7.5
python-decouple==3.1
pytz==2018.7
redis==2.10.3
s3transfer==0.1.13
six==1.11.0
urllib3==1.24.1
